rename sub-task-agent -> sub-coordinating-agent. better code organization / splitting up
