+++ apps/hash-ai-worker-ts/src/activities/flow-activities/research-entities-action.ai.test.ts
@@ -1,4 +1,4 @@
-import "../../../shared/testing-utilities/mock-get-flow-context.js";
+import "../../shared/testing-utilities/mock-get-flow-context.js";
 
 import {
   existsSync,
@@ -12,8 +12,8 @@ import { fileURLToPath } from "node:url";
 
 import { expect, test } from "vitest";
 
-import { researchEntitiesAction } from "../research-entities-action.js";
-import type { CoordinatingAgentState } from "./coordinating-agent.js";
+import { researchEntitiesAction } from "./research-entities-action.js";
+import type { CoordinatingAgentState } from "./research-entities-action/shared/coordinators.js";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
