+++ apps/hash-ai-worker-ts/src/activities/flow-activities/research-entities-action/shared/deduplicate-entities.ts
@@ -1,20 +1,20 @@
 import type { EntityId } from "@local/hash-graph-types/entity";
 import dedent from "dedent";
 
-import { logger } from "../../shared/activity-logger.js";
-import { getFlowContext } from "../../shared/get-flow-context.js";
-import { getLlmResponse } from "../../shared/get-llm-response.js";
-import type { AnthropicMessageModel } from "../../shared/get-llm-response/anthropic-client.js";
-import { getToolCallsFromLlmAssistantMessage } from "../../shared/get-llm-response/llm-message.js";
+import { logger } from "../../../shared/activity-logger.js";
+import { getFlowContext } from "../../../shared/get-flow-context.js";
+import { getLlmResponse } from "../../../shared/get-llm-response.js";
+import type { AnthropicMessageModel } from "../../../shared/get-llm-response/anthropic-client.js";
+import { getToolCallsFromLlmAssistantMessage } from "../../../shared/get-llm-response/llm-message.js";
 import type {
   LlmParams,
   LlmToolDefinition,
   LlmUsage,
-} from "../../shared/get-llm-response/types.js";
-import { graphApiClient } from "../../shared/graph-api-client.js";
-import type { PermittedOpenAiModel } from "../../shared/openai-client.js";
-import type { LocalEntitySummary } from "../shared/infer-summaries-then-claims-from-text/get-entity-summaries-from-text.js";
-import type { ExistingEntitySummary } from "./summarize-existing-entities.js";
+} from "../../../shared/get-llm-response/types.js";
+import { graphApiClient } from "../../../shared/graph-api-client.js";
+import type { PermittedOpenAiModel } from "../../../shared/openai-client.js";
+import type { LocalEntitySummary } from "../../shared/infer-summaries-then-claims-from-text/get-entity-summaries-from-text.js";
+import type { ExistingEntitySummary } from "../coordinating-agent/summarize-existing-entities.js";
 
 /**
  * @todo
