+++ apps/hash-ai-worker-ts/src/activities/flow-activities/research-entities-action/sub-coordinating-agent/input.ts
@@ -1,0 +1,10 @@
+import type { DereferencedEntityType } from "../../../shared/dereference-entity-type.js";
+import type { LocalEntitySummary } from "../../shared/infer-summaries-then-claims-from-text/get-entity-summaries-from-text.js";
+import type { Claim } from "../../shared/infer-summaries-then-claims-from-text/types.js";
+
+export type SubCoordinatingAgentInput = {
+  goal: string;
+  relevantEntities: LocalEntitySummary[];
+  existingClaimsAboutRelevantEntities: Claim[];
+  entityTypes: DereferencedEntityType[];
+};
