+++ apps/hash-ai-worker-ts/src/activities/flow-activities/research-entities-action/sub-coordinating-agent/state.ts
@@ -1,0 +1,18 @@
+import type { CompletedCoordinatorToolCall } from "../shared/coordinator-tools.js";
+import type { CoordinatingAgentState } from "../shared/coordinators.js";
+
+import { SubCoordinatingAgentToolName } from "./sub-coordinator-tools.js";
+
+export type SubCoordinatingAgentState = Pick<
+  CoordinatingAgentState,
+  | "plan"
+  | "entitySummaries"
+  | "inferredClaims"
+  | "resourcesNotVisited"
+  | "resourceUrlsVisited"
+  | "webQueriesMade"
+> & {
+  previousCalls: {
+    completedToolCalls: CompletedCoordinatorToolCall<SubCoordinatingAgentToolName>[];
+  }[];
+};
