+++ apps/hash-ai-worker-ts/src/activities/flow-activities/shared/propose-entities-from-claims.ts
@@ -10,7 +10,7 @@ import type { DereferencedEntityType } from "../../shared/dereference-entity-typ
 import { getFlowContext } from "../../shared/get-flow-context.js";
 import { logProgress } from "../../shared/log-progress.js";
 import { stringify } from "../../shared/stringify.js";
-import type { ExistingEntitySummary } from "../research-entities-action/summarize-existing-entities.js";
+import type { ExistingEntitySummary } from "../research-entities-action/coordinating-agent/summarize-existing-entities.js";
 import type { LocalEntitySummary } from "./infer-summaries-then-claims-from-text/get-entity-summaries-from-text.js";
 import type { Claim } from "./infer-summaries-then-claims-from-text/types.js";
 import { proposeEntityFromClaimsAgent } from "./propose-entities-from-claims/propose-entity-from-claims-agent.js";
